feature: adding some error handling for create update and delete. Closes #10, #21 and #22.

+++ src/application/usecases/delete-flight-travel.usecase.ts
@@ -1,3 +1,4 @@
+import { FlightTravelNotFound } from "../exceptions/flight-travel.exceptions";
 import { FlightTravelRepository } from "../flight-travel.repository";
 import { Injectable } from "@nestjs/common";
 
@@ -6,6 +7,9 @@ export class DeleteFlightTravelUseCase {
   constructor(private readonly flightTravelRepository: FlightTravelRepository) { }
 
   async handle(id: number): Promise<void> {
+    const flightTravelById = await this.flightTravelRepository.getById(id);
+    if (!flightTravelById) throw new FlightTravelNotFound();
+
     await this.flightTravelRepository.deleteById(id);
   }
 }
